requirements.txt
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-mgmt-compute==30.5.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.0.1
azure-mgmt-sql==3.0.1
azure-mgmt-storage==21.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
idna==3.6
isodate==0.6.1
loguru==0.7.2
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
